requirements.txt
Flask
pyTelegramBotAPI
python-dotenv
requests